+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Pipeline input for Disconnect-Service.
+
 ## [1.1.0] - 2022-07-30
 
 ### Added